refactor(disount): create hook to handle discount form logic
